refactor(kubeadm): remove deprecated command `alpha kubeconfig`
