+++ cmd/kubeadm/app/cmd/alpha/alpha.go
@@ -29,19 +29,5 @@ func NewCmdAlpha(in io.Reader, out io.Writer) *cobra.Command {
 		Short: "Kubeadm experimental sub-commands",
 	}
 
-	kubeconfigCmd := NewCmdKubeConfigUtility(out)
-	deprecateCommand(`please use the same command under "kubeadm kubeconfig"`, kubeconfigCmd)
-	cmd.AddCommand(kubeconfigCmd)
-
 	return cmd
-}
-
-func deprecateCommand(msg string, cmds ...*cobra.Command) {
-	for _, cmd := range cmds {
-		cmd.Deprecated = msg
-		childCmds := cmd.Commands()
-		if len(childCmds) > 0 {
-			deprecateCommand(msg, childCmds...)
-		}
-	}
 }
